Game: Store EM pointer in Object
This allows EM initialization to be done in the constructor, where it
belongs. It also allows for EM allocation to apply alignment. (Although
the possibility of Code::Word having stricter alignment than Object is
remote, to say the least.) There is also a small speedup for accessing
EMs, by avoiding a virtual call.